+++ tasks/ulshow.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+# Puppet Task Name: ulshow
+
+PATH=/usr/local/sbin:/sbin:/bin:/usr/sbin:/usr/bin:/opt/puppetlabs/bin:/usr/local/bin/rtpproxy:/root/bin:/usr/local/bin:/opt/puppetlabs/bin/=
+opensipsctl ul show $PT_opts $PT_user