+++ CHANGELOG.md
@@ -2,6 +2,7 @@ v0.4.x - 2019-09-02 Antonio Alisio de Meneses Cordeiro <alisio.meneses@gmail.com
 
 - Fix ineffective ratelimit function
 - Change default rate limit to 10
+- Add ul show Task
 
 v0.4.4 - 2019-09-02 Antonio Alisio de Meneses Cordeiro <alisio.meneses@gmail.com>